pyproject.toml
[tool.poetry]
name = "holidays"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
notebook = "^7.2.2"
gitpython = "^3.1.43"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
